Completely re-write the delay loop and the calibration routine
While the previous method worked fine on DOSBox with the pentium_slow
setting, it completely failed on a 4.77MHz 8088 (e.g., Tandy
1000HX). The new algorithm uses an interpolation method similar to a
Bresenham run-slice line drawing algorithm. This avoids mulitplication
and division in the time-critical code, and it maintains a decent
level of accuracy.

The calibration routine should also be a bit more forgiving of wildly
different CPU speeds.