Completely rewrite delay loop and delay loop calibration
The new algorithm is much, much more accurate. If finds the correct
delay parameters using a modified binary search to home in on the ideal
parameters. The unfortunate consequence is that it takes much longer on
fast PCs.

In addition, wait_44khz was modified to use 16-bit arithmetic. This will
perform much better on slow CPUs. The improved performance will
translate to more accurate timing.

The calculated values on my Tandy 1000HX are good, but they are still
suboptimal. The calculated parameters are n = 27000, d = 26999. I have
found that n = 27000, d = 23895 produce much better results.