Add user settable PSG base IO address
Tandy 1000RLX and Tandy 1000RSX are AT-class machines, so port C0 is
used by the second DMA controller. Tandy moved the PSG to port 1E0 on
these systems. The PicoGUS and the old lo-tech.co.uk Tandy Sound Card
default to using port 2C0. This adds a command line option to set the IO
base address.

It may yet be useful to detect AT systems and require the user to
specify a base address (since it cannot be the default C0).

Tested with PicoGUS set to IO port 2C0.